jsdialog: change sendDialogEvent API parameter
WindowId with 64bit length is needed

+++ desktop/source/lib/init.cxx
@@ -1014,7 +1014,7 @@ static void doc_removeTextContext(LibreOfficeKitDocument* pThis,
                                   int nCharBefore,
                                   int nCharAfter);
 static void doc_sendDialogEvent(LibreOfficeKitDocument* pThis,
-                               unsigned nLOKWindowId,
+                               unsigned long long int nLOKWindowId,
                                const char* pArguments);
 static void doc_postWindowKeyEvent(LibreOfficeKitDocument* pThis,
                                    unsigned nLOKWindowId,
@@ -3578,7 +3578,7 @@ public:
     virtual void SAL_CALL disposing(const css::lang::EventObject&) override {}
 };
 
-static void doc_sendDialogEvent(LibreOfficeKitDocument* /*pThis*/, unsigned nWindowId, const char* pArguments)
+static void doc_sendDialogEvent(LibreOfficeKitDocument* /*pThis*/, unsigned long long int nWindowId, const char* pArguments)
 {
     SolarMutexGuard aGuard;
